langchain-assistant/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf